Ignore bits that were introduced as padding in `transformUnitSize` (fixes #40)
Base64 embeds bytes into 6 bit units. It takes two units to properly encode a single byte. If you then decode this, you can conclude that the 4 additional bits contain no information. (Unless the input violates our assumption that it is a base64 encoding.)

You can conclude that if $y = \lceil xs_{input} / s_{output}\rceil$ with x an integer, then $x <= \lfloor ys_{output} / s_{input}\rfloor$. This commit checks every added unit if it is the last, and doesn't add the null byte if it is. It might be more efficient to just use the larger array and slice off the null byte when it exists.

It can be impossible to determine a true zero byte from padding if `inputSize < outputSize`, but for binary-to-text that requires an alphabet of more than 256 characters.

Signed-off-by: anderium <33520919+anderium@users.noreply.github.com>